+++ packages/documentation/examples/drag-handles.tsx
@@ -1,0 +1,3 @@
+import { AllDragHandleVariants } from './guidelines/all-drag-handle-variants';
+
+export default AllDragHandleVariants;