+++ packages/documentation/examples/guidelines/entire-entity-is-draggable-with-grouped-items.tsx
@@ -5,8 +5,8 @@ import invariant from 'tiny-invariant';
 
 import { IconButton } from '@atlaskit/button/new';
 import DropdownMenu, { DropdownItem, DropdownItemGroup } from '@atlaskit/dropdown-menu';
+import DragHandleVerticalIcon from '@atlaskit/icon/core/drag-handle-vertical';
 import ChevronRightIcon from '@atlaskit/icon/core/migration/chevron-right';
-import DragHandleVerticalIcon from '@atlaskit/icon/core/migration/drag-handle-vertical--drag-handler';
 import EditorMoreIcon from '@atlaskit/icon/core/migration/show-more-horizontal--editor-more';
 import { fg } from '@atlaskit/platform-feature-flags';
 import { draggable } from '@atlaskit/pragmatic-drag-and-drop/element/adapter';