+++ packages/documentation/examples/pieces/subtasks/primitives/notion/subtask.tsx
@@ -7,7 +7,7 @@ import { forwardRef, type HTMLAttributes, memo, type ReactNode, type Ref } from 
 // eslint-disable-next-line @atlaskit/ui-styling-standard/use-compiled -- Ignored via go/DSP-18766
 import { css, jsx, type SerializedStyles } from '@emotion/react';
 
-import DragHandleVerticalIcon from '@atlaskit/icon/core/migration/drag-handle-vertical--drag-handler';
+import DragHandleVerticalIcon from '@atlaskit/icon/core/drag-handle-vertical';
 import Lozenge from '@atlaskit/lozenge';
 import { token } from '@atlaskit/tokens';